PKG -- [fcl] update default-config
